pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "salesforce-gift-eval"
dynamic = [
  "version",
]
dependencies = [
  "pandas==2.0.0",
  "gluonts~=0.15.1",
  "numpy~=1.26.0",
  "einops==0.7.*",
  "python-dotenv==1.0.0",
  "hydra-core==1.3",
  "datasets~=2.17.1",
  "orjson",
  "matplotlib~=3.9.2",
  "tsfeatures",
  "ray",
  "scipy~=1.11.3",
]
requires-python = ">=3.10"
authors = [
  {name = "Taha Aksu", email = "iaksu@salesforce.com"},
  {name = "Gerald Woo", email = "gwoo@salesforce.com"},
  {name = "Chenghao Liu", email = "chenghao.liu@salesforce.com"},
  {name = "Juncheng Liu", email = "juncheng.liu@salesforce.com"},
  {name = "Xu Liu", email = "xu.liu@salesforce.com"},
  {name = "Caiming Xiong"},
  {name = "Silvio Savarese"},
  {name = "Doyen Sahoo"},
]
maintainers = [
  {name = "Taha Aksu", email = "iaksu@salesforce.com"},
  {name = "Chenghao Liu", email = "chenghao.liu@salesforce.com"},
]
description = "GIFT-Eval Benchmark"
#readme = "README.md"
#license = {file = "LICENSE.txt"}
keywords = ["Time Series Forecasting", "Transformer", "Deep Learning", "PyTorch"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.optional-dependencies]
baseline = [
  "torch>=2.1",
  "statsforecast",
  "tensorboard",
  "torch>=2.1",
  "lightning>=2.0",
]

[tool.hatch.version]
path = "src/gift_eval/__about__.py"

[tool.hatch.build]
packages = ["src/gift_eval"]

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]
